requirements.txt
Django==5.2.7
psycopg2-binary==2.9.10
python-decouple==3.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
requests==2.31.0
